+++ functions/Handlers/index.js
@@ -1,0 +1,6 @@
+exports.signup = require('./SignupHandler').handler;
+exports.createclass = require('./CreateClassHandler').handler;
+exports.joinclass = require('./JoinClassHandler').handler;
+exports.getclasses = require('./GetClassesHandler').handler;
+exports.requestclass = require('./RequestClassHandler').handler;
+exports.approveclass = require('./ApproveClassHandler').handler;
